Add warnings to Engine, make glad actual target
Because including source files in the interface means that they get
added to dependant targets. That is:

my_target (-Wall, -Werror, -pedantic):
- file1.cpp

my_other_target (IMPORTED) (-Wno-whatever):
- file2.cpp

In this example, file2.cpp gets added to my_target, and it gets
the compile options of my_target. That means it needs to be compiled
in its own context so that it can have its own compile options and
such.