Fixed M1 Glfw SigAbort when creating a window
Due to weirdness in the macOs system, when attempting to create a window it will return 0 extensions, which we had labeled as an error earlier. Now that we have fixed this issue we also ran into the problem of adding window attributes mac did not like. Added an ifdef and a nullptr check to solve the issues.
NOTE:: THIS DOES NOT SOLVE THE PROBLEM OF CONAN NOT PROVIDING A PROPER VULKAN LOADER FOR MACOS!!