Added mechanism to type pun semi-safe
this handles most cases, as long as C++ 17 is used and the macros don't cry.

in C++ 20 std::bit_set is used.
some compilers support it in earlier C++ versions as a sort of compiler intrinsic, if one is available it might use it.
if the compiler has not implemented std::bit_set yet or a built in implementation cannot be found, it will fallback on memcpy for GNU based compilers except clang and icc, these use reinterpret_cast as fallback.

the list of gnu exceptions is a tricky one to pin down and this is my current best attempt.